[level 0] Title: 주사위의 개수, Time: 0.03 ms, Memory: 73.7 MB -BaekjoonHub

+++ 프로그래머스/lv0/120845. 주사위의 개수/주사위의 개수.java
@@ -1,5 +1,5 @@
 class Solution {
     public int solution(int[] box, int n) {
-        return (int)(box[0] / n) * (int)(box[1] / n) * (int)(box[2] / n);
+        return (box[0] / n) * (box[1] / n) * (box[2] / n);
     }
 }
